fsm based on ao implementation.

+++ components/ao_core/include/ao_core.h
@@ -89,7 +89,7 @@ esp_err_t ao_start(ao_t* self, UBaseType_t prio, uint32_t stack_words);
  * @return ESP_OK if the event is posted successfully, or an error code otherwise.
  * @note The active object must be started using ao_start() before posting events to it.
  */
-esp_err_t ao_post(ao_t* self, uint16_t type, const void* payload, uint16_t len, TickType_t to_ticks);
+esp_err_t ao_post(ao_t* self, uint8_t type, const void* payload, uint8_t len, TickType_t to_ticks);
 
 /**
  * @brief Stops the active object.
